+++ src/Network/Network/AzureFirewall/NetworkRule/NewAzureFirewallNetworkRuleCommand.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.Network.Models;
+using System.Text.RegularExpressions;
 using MNM = Microsoft.Azure.Management.Network.Models;
 
 namespace Microsoft.Azure.Commands.Network
@@ -43,10 +44,16 @@ namespace Microsoft.Azure.Commands.Network
         public string[] SourceAddress { get; set; }
 
         [Parameter(
-            Mandatory = true,
+            Mandatory = false,
             HelpMessage = "The destination addresses of the rule")]
         [ValidateNotNullOrEmpty]
         public string[] DestinationAddress { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "The destination FQDNs of the rule")]
+        [ValidateNotNullOrEmpty]
+        public string[] DestinationFqdns { get; set; }
 
         [Parameter(
             Mandatory = true,
@@ -68,7 +75,27 @@ namespace Microsoft.Azure.Commands.Network
         public override void Execute()
         {
             base.Execute();
-            
+
+            if (DestinationFqdns != null)
+            {
+                foreach (string fqdn in DestinationFqdns)
+                {
+                    ValidateIsFqdn(fqdn);
+                }
+            }
+
+            // Only one of DestinationAddress or DestinationFqdns is allowed
+            if ((DestinationAddress != null) && (DestinationFqdns != null))
+            {
+                throw new ArgumentException("Both DestinationAddress and DestinationFqdns not allowed");
+            }
+
+            // One of DestinationAddress or DestinationFqdns must be present
+            if ((DestinationAddress == null) && (DestinationFqdns == null))
+            {
+                throw new ArgumentException("Either DestinationAddress or DestinationFqdns is required");
+            }
+
             var networkRule = new PSAzureFirewallNetworkRule
             {
                 Name = this.Name,
@@ -76,9 +103,20 @@ namespace Microsoft.Azure.Commands.Network
                 Protocols = this.Protocol?.ToList(),
                 SourceAddresses = this.SourceAddress?.ToList(),
                 DestinationAddresses = this.DestinationAddress?.ToList(),
+                DestinationFqdns = this.DestinationFqdns?.ToList(),
                 DestinationPorts = this.DestinationPort?.ToList()
             };
             WriteObject(networkRule);
+        }
+
+        private void ValidateIsFqdn(string fqdn)
+        {
+            var fqdnRegEx = new Regex("^\\*$|^[a-zA-Z0-9]+(([a-zA-Z0-9_\\-]*[a-zA-Z0-9]+)*\\.)*(?:[a-zA-Z0-9]{2,})$");
+
+            if (!fqdnRegEx.IsMatch(fqdn))
+            {
+                throw new ArgumentException($"Invalid value {fqdn}.");
+            }
         }
     }
 }